requirements.txt
strictyaml==1.0.0
attrs==19.1.0
